MicLoopBack/requirements.txt
cffi==2.0.0
numpy==2.3.3
pycparser==2.23
sounddevice==0.5.2
websocket-client==1.8.0
